+++ Source/FlightBrowse/api/Airline.cpp
@@ -1,0 +1,12 @@
+#include "Airline.h"
+#include "Flight.h"
+
+Airline::Airline() 
+{
+
+};
+
+Flight** Airline::getFlights()
+{
+
+};
