+++ Source/FlightBrowse/api/Airline.h
@@ -1,0 +1,12 @@
+#include "Flight.h"
+
+class Airline
+{
+private:
+    Flight** flights;
+
+public:
+    Airline();
+
+    Flight** getFlights();
+};
